pyproject.toml
[project]
name = "seismic-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "datime>=2025.0",
    "httpx>=0.28.1",
    "notch>=2025.2",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "ruff>=0.14.0",
]

[tool.ruff.lint]
select = ["ANN", "E", "F", "FURB", "I", "PERF", "RUF", "S", "UP"]
